requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==1.5.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
decorator==4.1.2
enum34==1.1.6
Flask==0.12.2
funcsigs==1.0.2
h5py==2.7.1
html5lib==0.9999999
idna==2.6
ipdb==0.10.3
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
Keras==2.0.8
Markdown==2.6.9
MarkupSafe==1.0
mock==2.0.0
mysql==0.0.1
MySQL-python==1.2.5
numpy==1.13.3
olefile==0.44
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
py==1.4.34
Pygments==2.2.0
pytest==3.2.3
PyYAML==3.12
requests==2.18.4
scandir==1.6
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.1.14
tensorflow==1.3.0
tensorflow-tensorboard==0.1.8
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
